requirements.txt
requests==2.31.0
langchain==0.0.295
python-dotenv==1.0.0
pathlib
ibm-watson==7.0.0
ibm-generative-ai==0.2.8
numpy
toml==0.10.2
pymysql
ibm-watson-machine-learning==1.0.312
huggingface_hub
unstructured
streamlit
streamlit-chat
lark
xlsxwriter==3.1.4
pypandoc==1.11
pymupdf
chromadb==0.4.14
minio